Module proxy_wasm_experimental::hostcalls

Constants

NO_BODY
Represents empty body.

NO_HEADERS
Represents empty headers map.

NO_TRAILERS
Represents empty trailers map.

Functions

add_map_value
Adds a key-value pair to a given map.

close_stream
Terminates processing of a given stream, i.e. HTTP request or HTTP response.

continue_stream
Resumes processing of a given stream, i.e. HTTP request or HTTP response.

define_metric

dequeue_shared_queue
Returns data from the end of a given queue.

dispatch_http_call
Dispatches an HTTP call to a given upstream.

done
Indicates to the host environment that Wasm VM side is done processing current context.

enqueue_shared_queue
Adds a value to the front of a given queue.

get_buffer
Returns content from a given buffer.

get_current_time
Returns current system time.

get_map
Returns all key-value pairs from a given map.

get_map_value
Returns value of a given key from a given map.

get_metric

get_property
Returns value of a property in the current context.

get_shared_data
Returns shared data by key.

increment_metric

log
Logs a message at a given log level.

record_metric

register_shared_queue
Registers a shared queue with a given name.

resolve_shared_queue
Looks up for an existing shared queue with a given name.

send_http_response
Sends HTTP response without forwarding request to the upstream.

set_buffer
Mutates content in a given buffer.

set_effective_context
Changes the effective context.

set_map
Sets all key-value pairs in a given map.

set_map_value
Sets / replaces / removes value of given key from a given map.

set_property
Sets property to a given value in the current context.

set_shared_data
Sets shared data by key.

set_tick_period
Sets the timer to a given period.